requirements.txt
Flask
Flask-SQLAlchemy
Flask-MySQLdb
subprocess.run